Extensions were divided by types.
