+++ cs/TagsCloudVisualization/Extensions/SizeExtensions.cs
@@ -1,0 +1,12 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.Extensions
+{
+    public static class SizeExtensions
+    {
+        public static Size Multiply(this Size size, double number)
+        {
+            return new Size((int)(size.Width * number), (int)(size.Height * number));
+        }
+    }
+}
